refactor: /presigned 공통부분 RequestMapping으로 빼기
