+++ src/main/java/ojosama/talkak/video/controller/AwsS3Controller.java
@@ -1,8 +1,8 @@
 package ojosama.talkak.video.controller;
 
 import lombok.RequiredArgsConstructor;
-import ojosama.talkak.video.dto.AwsS3RequestDto;
-import ojosama.talkak.video.dto.AwsS3ResponseDto;
+import ojosama.talkak.video.dto.AwsS3Request;
+import ojosama.talkak.video.dto.AwsS3Response;
 import ojosama.talkak.video.service.AwsS3Service;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -10,24 +10,24 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api/files")
+@RequestMapping("/api/files/presigned")
 @RequiredArgsConstructor
 public class AwsS3Controller {
 
     private final AwsS3Service awsS3Service;
 
-    @PostMapping("/presigned/upload")
-    public AwsS3ResponseDto getPresignedUrlToUpload(@RequestBody AwsS3RequestDto awsS3RequestDto) {
-        return awsS3Service.getPresignedUrlToUpload(awsS3RequestDto);
+    @PostMapping("/upload")
+    public AwsS3Response getPresignedUrlToUpload(@RequestBody AwsS3Request awsS3Request) {
+        return awsS3Service.getPresignedUrlToUpload(awsS3Request);
     }
 
-    @PostMapping("/presigned/download")
-    public AwsS3ResponseDto getPresignedUrlToDownload(@RequestBody AwsS3RequestDto awsS3RequestDto) {
-        return awsS3Service.getPresignedUrlToDownload(awsS3RequestDto);
+    @PostMapping("/download")
+    public AwsS3Response getPresignedUrlToDownload(@RequestBody AwsS3Request awsS3Request) {
+        return awsS3Service.getPresignedUrlToDownload(awsS3Request);
     }
 
-    @PostMapping("/presigned/delete")
-    public AwsS3ResponseDto getPresignedUrlToDelete(@RequestBody AwsS3RequestDto awsS3RequestDto) {
-        return awsS3Service.getPresignedUrlToDelete(awsS3RequestDto);
+    @PostMapping("/delete")
+    public AwsS3Response getPresignedUrlToDelete(@RequestBody AwsS3Request awsS3Request) {
+        return awsS3Service.getPresignedUrlToDelete(awsS3Request);
     }
 }
